[S11P21D109-294] - 입찰 응답과 경매 로그 반환 타입이 일치 되도록 수정

See merge request s11-blockchain-nft-sub1/S11P21D109!144

+++ backend/spring-server/src/main/java/com/semonemo/spring_server/domain/auction/dto/request/BidRequestDTO.java
@@ -9,5 +9,5 @@ import lombok.ToString;
 @ToString
 public class BidRequestDTO {
 	private long userId;
-	private int bidAmount;
+	private long bidAmount;
 }
